chore(system): extract the worker to its own module
