+++ core/src/test/java/io/kestra/core/runners/TestMethodScopedWorker.java
@@ -3,10 +3,10 @@ package io.kestra.core.runners;
 import io.kestra.core.server.ServiceStateChangeEvent;
 import io.kestra.core.services.WorkerGroupService;
 import io.kestra.core.utils.ExecutorsUtils;
+import io.kestra.worker.DefaultWorker;
 import io.micronaut.context.annotation.Parameter;
 import io.micronaut.context.annotation.Prototype;
 import io.micronaut.context.event.ApplicationEventPublisher;
-import io.micronaut.core.annotation.Introspected;
 import io.micronaut.core.annotation.Nullable;
 import jakarta.inject.Inject;
 
@@ -14,8 +14,7 @@ import jakarta.inject.Inject;
  * This worker is a special worker which won't close every queue allowing it to be ran and closed within a test without
  * preventing the Micronaut context to be used for further tests with queues still being up
  */
-@Prototype
-public class TestMethodScopedWorker extends Worker {
+public class TestMethodScopedWorker extends DefaultWorker {
     @Inject
     public TestMethodScopedWorker(@Parameter String workerId,
                                   @Parameter Integer numThreads,
