+++ webserver/build.gradle
@@ -34,6 +34,7 @@ dependencies {
     testAnnotationProcessor project(':processor')
     testImplementation project(':core').sourceSets.test.output
     testImplementation project(':storage-local')
+    testImplementation project(':worker')
     testImplementation "org.wiremock:wiremock-jetty12"
     testImplementation "org.awaitility:awaitility"
