+++ webserver/src/test/java/io/kestra/webserver/controllers/api/ClusterControllerTest.java
@@ -1,9 +1,9 @@
 package io.kestra.webserver.controllers.api;
 
 import io.kestra.core.junit.annotations.KestraTest;
-import io.kestra.core.runners.Worker;
 import io.kestra.core.server.ServerInstance;
 import io.kestra.core.server.ServiceInstance;
+import io.kestra.worker.DefaultWorker;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.client.annotation.Client;
 import io.micronaut.reactor.http.client.ReactorHttpClient;
@@ -19,7 +19,7 @@ class ClusterControllerTest {
     ReactorHttpClient client;
 
     @Inject
-    Worker worker;
+    DefaultWorker worker;
 
     @Test
     void shouldGetServiceInfo() {
