+++ worker/src/main/java/io/kestra/worker/WorkerSecurityService.java
@@ -1,4 +1,4 @@
-package io.kestra.core.runners;
+package io.kestra.worker;
 
 import io.kestra.core.models.flows.State;
 import jakarta.inject.Singleton;
