+++ worker/src/main/java/io/kestra/worker/WorkerTriggerCallable.java
@@ -1,8 +1,10 @@
-package io.kestra.core.runners;
+package io.kestra.worker;
 
 import io.kestra.core.models.executions.Execution;
 import io.kestra.core.models.flows.State;
 import io.kestra.core.models.triggers.PollingTriggerInterface;
+import io.kestra.core.runners.RunContext;
+import io.kestra.core.runners.WorkerTrigger;
 import lombok.Getter;
 
 import java.util.Optional;
